idをintからuintに修正

+++ src/entity/model/chair/id.go
@@ -4,11 +4,11 @@ import "errors"
 
 type ID uint
 
-func NewChairID(val int) (*ID, error) {
+func NewChairID(val uint) (*ID, error) {
 	if val < 1 {
 		return nil, errors.New("idは1以上の整数である必要があります。")
 	}
-	id := ID(uint(val))
+	id := ID(val)
 	return &id, nil
 }
